dynamic side panel and text sizing
The homepage now has a more dynamic side panel that displays when the screen is big enough but only displays the drawer when it's not big enough.
The animation timers now only run when the component is mounted, because it was causing issues if it was destroyed from scrolling away from it.

I created a mostly custom font size getter so that the font size changes with bigger screens. it doesn't work super well right now as the animated text stays the same even if the window changes sizes in real time. The web & mobile developer text however does, because that is just a text wrapped in automatic opacity.

The side panel also has some icons and is styled a bit more so that it doesn't look so weird.