Projects on hover and video player
the projects section now has images to show what the project looks like, as well as a video (only for the rock canyon demo for now) player popup to let the user see what the web app actually interacts like.

I have a nice simple onhover effect going on for each card, and the ontap is on the project card as a whole and not just the image.