Icons on About Cards
I didn't like how the text was looking on the about cards so I made the decision to use Icons on the cards, with a bigger title to display the things I'm "about."
I created a custom SVG for Project Management, found a flutter SVG, and modified the color of the "onHover" so that there is some opacity as well.
I LOVE how it's turned out so far. I still have to finish the about cards but this is a good step to making the site prettier as well as functional.