Chadlyalan/issue12 (#14)
* Icons on About Cards

I didn't like how the text was looking on the about cards so I made the decision to use Icons on the cards, with a bigger title to display the things I'm "about."
I created a custom SVG for Project Management, found a flutter SVG, and modified the color of the "onHover" so that there is some opacity as well.
I LOVE how it's turned out so far. I still have to finish the about cards but this is a good step to making the site prettier as well as functional.

* about me cleaning

tidying up the about me section a bit more by adding some more logo/icons for the sections. Now they're a bit more colorful. I have added a couple other's that I plan to use in other places on the website, like the github logo.