Chadlyalan/issue13 (#15)
* Icons on About Cards

I didn't like how the text was looking on the about cards so I made the decision to use Icons on the cards, with a bigger title to display the things I'm "about."
I created a custom SVG for Project Management, found a flutter SVG, and modified the color of the "onHover" so that there is some opacity as well.
I LOVE how it's turned out so far. I still have to finish the about cards but this is a good step to making the site prettier as well as functional.

* about me cleaning

tidying up the about me section a bit more by adding some more logo/icons for the sections. Now they're a bit more colorful. I have added a couple other's that I plan to use in other places on the website, like the github logo.

* project size + hyperlink

The project's page size is now dynamic to how many cards there are rather than a single page's height. I can add a couple other projects to fill up the space.

There is currently a hardcoded button(link) to the founders pass demo site.

* link logic

the hyper link in the project card now only displays if there is a link sent to the project card. Otherwise it shows an empty container.

* socials and const fixing

I've changed the "contact me" section to "socials" which will only exist on smaller screens. On bigger screens I think I'm going to show them smallish on the side panel.