cleaning up socials
I'm going to give up on showing the resume in a PDF for now. I don't think most people are going to want to see the pdf while they're at my website anyway so it will be a feature I add later once the website is up and polished. for now socials are good where they're at.