solved - for now
The issue was with how i was scrolling to my widget's in the tree.
the scrollTo function makes a calculation as to where that widget will be, but they haven't been rendered yet so it really doesn't know where it is going to be at. This causes some confusion.
There were two solutions:
1. make each page the same (or at least static) size.
2. Use jumpTo rather than scrollTo

I went with jumpTo. This doesn't have a nice scrolling animation but it does get the job done for now. it's not my favorite thing but it is fixed and I have to move on for now. I'll close this issue for now as it is fixed.