rock-garden project logo (#26)
I fixed the rock garden project (for redpoint) with the rock garden logo that we used in redpoint. I might swap it out for a real rockgarden logo at some point but i kinda dig this one.

For the light theme I'm going to have ot use a different one? or at least I'll have to change the lines from white to black.
I am thinking I'll have to find something more crisp to use for the other projects as well..