About card title font size
Instead of removing the title all together when the screen gets smaller, I decided I still want the title for the about cards, so I'm adding a little width check to decide what the minimum font size for the title should be. This way the font is small for small screens, and big for bigger screens.