project section content update
I've replaced the react project's logo with a weather logo to reflect what the project is actually about.

I fixed the font in the project section description to match the smaller one in the About me section as well.

I've added a prop for the project card to decide which video to display, now the weather react.js project has it's own little demo video as well.

From here the only project missing a demo video is the redpoint (authoring tools) for rock garden. It will be easy to add the video once I take a recording of it.