videoname parameter and no video message
The project cards now accept an optional parameter called "videoName" that is the asset path in the form of a string of the video to render for the user onClick.
If the video name isn't there, when the card is clicked on, we will display an error message stating "sorry there is no video demo yet".
This is temporary as we will have a video for each card at some point.  But this should clear up some errors and issues for now.