Chadlyalan/issue20 (#32)
* project section content update

I've replaced the react project's logo with a weather logo to reflect what the project is actually about.

I fixed the font in the project section description to match the smaller one in the About me section as well.

I've added a prop for the project card to decide which video to display, now the weather react.js project has it's own little demo video as well.

From here the only project missing a demo video is the redpoint (authoring tools) for rock garden. It will be easy to add the video once I take a recording of it.

* Project Card padding + firebase about me card

The error following the firebase card hasn't popped up in a while.

There is an uncaught error doing something weird with one of the project cards (the middle one in fact) that needs to be fixed.

* videoname parameter and no video message

The project cards now accept an optional parameter called "videoName" that is the asset path in the form of a string of the video to render for the user onClick.
If the video name isn't there, when the card is clicked on, we will display an error message stating "sorry there is no video demo yet".
This is temporary as we will have a video for each card at some point.  But this should clear up some errors and issues for now.