working animations
we converted this landing widget into a stateless widget because the animated text was freaking out when setState was called. I extracted all of the setState functionality to their individual widgets, and let landing page worry less about them.

Now the animations work just as they used to, AND when the theme get's toggled, they rebuild correctly.