text colors
these changes will update the text color for the light mode to white. White comes up fairly well on the tealish background, better than the black does at any rate. It's not the best, I'm not convinced this is the final color scheme, but it's a color scheme at least.