Chadlyalan/issue22 - Theme first draft (#33)
* working animations

we converted this landing widget into a stateless widget because the animated text was freaking out when setState was called. I extracted all of the setState functionality to their individual widgets, and let landing page worry less about them.

Now the animations work just as they used to, AND when the theme get's toggled, they rebuild correctly.

* Update landing.dart

converting to stateless widget

* a light theme

this likely won't be the final light theme scheme but it works for now. I'll continue playing around with some of the colors to see how I feel about them as time goes on.

Also a small grammatical fix to the About me section.

* text colors

these changes will update the text color for the light mode to white. White comes up fairly well on the tealish background, better than the black does at any rate. It's not the best, I'm not convinced this is the final color scheme, but it's a color scheme at least.