light theme overlap error fix
The colors aren't correct as those changes seem to be local on my PC but this includes the fix to the light theme's overlap issue. It was because I was using material3 on light, which I do like, but there was extra padding in the animated texts that was causing issues.

I also have fixed the appbar's color in this mode, as normal material had a plain white appbar in light mode by default.

I also realized that I didn't need the align widget in the landing page, as the column was already telling it's children where to be placed.