a light theme that works
maybe not the final product but it is a light theme that doesn't cause sizing issues when turning back and forth from light/dark.
I'm not sold on the orange cards but I kinda like them. I want to do something a little bit different with the project cards if I can think of what to do.