project card styling
Especially in the light theme, it felt like the project cards needed a little extra love. I added a color to the entire "card" to give it a little extra feeling on hover.

I haven't decided what I should do for the mobile site to make sure it feels good.