project card styling
I'm going to finish up the light theme mode changes with a project card update.

I've added some rounded borders to the project cards to hopefully give the impression that they're clickable from the get go. On mobile especially it's hard to tell that they're clickable so this should help a bit.