Chadlyalan/issue22 (#35)
* working animations

we converted this landing widget into a stateless widget because the animated text was freaking out when setState was called. I extracted all of the setState functionality to their individual widgets, and let landing page worry less about them.

Now the animations work just as they used to, AND when the theme get's toggled, they rebuild correctly.

* Update landing.dart

converting to stateless widget

* a light theme

this likely won't be the final light theme scheme but it works for now. I'll continue playing around with some of the colors to see how I feel about them as time goes on.

Also a small grammatical fix to the About me section.

* text colors

these changes will update the text color for the light mode to white. White comes up fairly well on the tealish background, better than the black does at any rate. It's not the best, I'm not convinced this is the final color scheme, but it's a color scheme at least.

* light theme overlap error fix

The colors aren't correct as those changes seem to be local on my PC but this includes the fix to the light theme's overlap issue. It was because I was using material3 on light, which I do like, but there was extra padding in the animated texts that was causing issues.

I also have fixed the appbar's color in this mode, as normal material had a plain white appbar in light mode by default.

I also realized that I didn't need the align widget in the landing page, as the column was already telling it's children where to be placed.

* a light theme that works

maybe not the final product but it is a light theme that doesn't cause sizing issues when turning back and forth from light/dark.
I'm not sold on the orange cards but I kinda like them. I want to do something a little bit different with the project cards if I can think of what to do.

* Update my_theme.dart

* Update my_theme.dart

* redpoint demo video

I cut together a demo video of redpoint finally. It should be working but has not been tested yet.

* project card styling

Especially in the light theme, it felt like the project cards needed a little extra love. I added a color to the entire "card" to give it a little extra feeling on hover.

I haven't decided what I should do for the mobile site to make sure it feels good.

* project card styling

I'm going to finish up the light theme mode changes with a project card update.

I've added some rounded borders to the project cards to hopefully give the impression that they're clickable from the get go. On mobile especially it's hard to tell that they're clickable so this should help a bit.

* project cleanup

adding some text to the project card for redpoint.

Changed the link button to say "go to site" rather than "a link" when there's a link to the live site of one of the project demo cards.